server/requirements.txt
cffi==1.15.1
pycparser==2.21
pydantic==1.10.6
PyNaCl==1.5.0
typing_extensions==4.5.0
websockets==10.4
